Split the council and chair scripts by concern
council-review.mjs mixed member definition and dispatch with diff
preparation and output assembly. chair-fallback.mjs mixed JSON recovery
with verdict rendering. Both files were also near the 300-line oxlint
budget, which a formatter would push them over.

Moves member parsing, the per-lens prompt, the request and the OpenRouter
fallback to council-members.mjs, and the JSON repair and verdict parse to
verdict-json.mjs. Pure moves: no logic, signature or output changes.

Both scripts' --selfcheck output is byte-identical before and after.
