+++ scripts/chair-fallback.mjs
@@ -16,6 +16,7 @@
 import fs from "node:fs";
 import { execFileSync } from "node:child_process";
 import { pathToFileURL } from "node:url";
+import { parseVerdict } from "./verdict-json.mjs";
 
 const MAX_DIFF_CHARS = 180_000;
 const TIMEOUT_MS = Number(process.env.CHAIR_FALLBACK_TIMEOUT_MS) || 180_000;
@@ -130,75 +131,6 @@ async function askChair(diff, council, diffTruncated) {
 // diff or a regex) and a raw newline or tab. Both are recoverable, and losing
 // an entire review to one stray backslash is exactly what a fallback exists to
 // prevent. Repairs only what is invalid; a well-formed reply is unchanged.
-const VALID_ESCAPE = new Set(['"', "\\", "/", "b", "f", "n", "r", "t", "u"]);
-const CONTROL_ESCAPE = { "\n": "\\n", "\r": "\\r", "\t": "\\t", "\b": "\\b", "\f": "\\f" };
-const HEX4 = /^[0-9a-fA-F]{4}/;
-// A backslash right before a quote is a genuine `\"` escape only if the string
-// keeps going afterward. If what follows is unambiguously the START of the
-// NEXT token — end of container (`}`/`]`), a bare `:` (this string was a key),
-// or `,` followed by another quoted key and `:` — that quote is the real
-// closing quote and the backslash is a stray one, e.g. a Windows path ending
-// in `\`. A bare `,` is NOT enough on its own: prose routinely quotes a word
-// and continues with a comma ("say \"hi\", and ..."), which is a `\"` mid-string,
-// not a closing quote, even though a comma follows it too.
-const STRUCTURAL_AFTER_STRING = /^\s*(?:[}\]]|:|,\s*"[^"\\]*"\s*:)/;
-
-function repairJsonStrings(text) {
-  let out = "";
-  let inString = false;
-  for (let i = 0; i < text.length; i++) {
-    const ch = text[i];
-    if (!inString) {
-      inString = ch === '"';
-      out += ch;
-    } else if (ch === '"') {
-      inString = false;
-      out += ch;
-    } else if (ch === "\\") {
-      const next = text[i + 1];
-      if (next === undefined) out += "\\\\";
-      else if (next === '"' && STRUCTURAL_AFTER_STRING.test(text.slice(i + 2))) {
-        out += "\\\\";
-      } else if (next === "u" && !HEX4.test(text.slice(i + 2, i + 6))) {
-        // `\u` is only a valid escape when 4 hex digits follow; otherwise it's
-        // a path like `C:\users`, and only the backslash is invalid.
-        out += "\\\\";
-      } else if (VALID_ESCAPE.has(next)) {
-        out += ch + next;
-        i++;
-      } else {
-        // Only escape the backslash; leave `next` for the next iteration so a
-        // control character right after an invalid escape still gets escaped
-        // instead of being copied into the string raw.
-        out += "\\\\";
-      }
-    } else if (ch < " ") {
-      out += CONTROL_ESCAPE[ch] || `\\u${ch.charCodeAt(0).toString(16).padStart(4, "0")}`;
-    } else out += ch;
-  }
-  return out;
-}
-
-// A model told to emit JSON still sometimes wraps it in a fence or prose. Take
-// the outermost braces rather than failing the whole review on formatting.
-function parseVerdict(text) {
-  const start = text.indexOf("{");
-  const end = text.lastIndexOf("}");
-  if (start === -1 || end <= start) throw new Error(`no JSON object in chair reply: ${truncate(text, 200)}`);
-  const candidate = text.slice(start, end + 1);
-  try {
-    return JSON.parse(candidate);
-  } catch (err) {
-    try {
-      return JSON.parse(repairJsonStrings(candidate));
-    } catch (repairErr) {
-      throw new Error(`chair reply is not JSON even after repair: ${err.message} (repair attempt: ${repairErr.message})`);
-    }
-  }
-}
-
-// One normalization, used by both the rendered body and the verdict. Two
-// call sites normalizing differently is how a parseable reply still throws.
 function normalizeFindings(parsed) {
   const raw = parsed?.findings;
   if (raw === undefined || raw === null) return [];
